Rewriting parser for the new RD site.
They did a redesign to make their site more sane, although it's still a
bit wonky.

The biggest change is there's an unofficial API I can query. Note that
it seems to hijack and rewrite requests through "hwakproxy". This is
slightly confusing, but hasn't caused any issues.

Annoyingly, I can only seem to get it to return a JSON object containing
a JS callback to which the whole HTML content of the page is passed. For
now, I strip away the characters surrounding the HTML, pass it to
BeautifulSoup, then extract the info I care about.

Overall, their use of an API is great for me. Retireving their entire
inventory is SUPER fast now, taking minutes instead of days. As with the
BJs parser, I sleep between each page query so as to not flood their
site.