Restructure inventory to store more info.
The inventory is now a list of JSON objects. Each object currently
contains the product's name and category. The category field is
structured to allow it to encode a single product in multiple
categories, and each category to have subcategories. While RD currently
has neither of these features, it brings it in line with my other store
inventory file. Also it's not unreasonable for RD to add this later.